backend/friends/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.8
python-dotenv
hvac
gunicorn
djangorestframework
PyJWT
certifi
requests
